Fix Installer Engine project being used in places it shouldn't be
