+++ src/ServiceControl.AcceptanceTests/Recoverability/When_a_native_integration_message_is_retried.cs
@@ -11,7 +11,6 @@
     using NServiceBus.Transport;
     using NUnit.Framework;
     using ServiceControl.MessageFailures.Api;
-    using ServiceControlInstaller.Engine.Instances;
     using TestSupport.EndpointTemplates;
     using Conventions = NServiceBus.AcceptanceTesting.Customization.Conventions;
 
@@ -36,16 +35,13 @@
                 .Run();
 
             // ASQ adds this value to it's message wrapper and writes the incoming message header
-            if (TransportIntegration.Name != TransportNames.AzureStorageQueue)
+            if (TransportIntegration.Name != "Azure Storage Queue")
             {
                 Assert.False(context.Headers.ContainsKey(Headers.MessageIntent), "Should not add the intent header");
             }
 
             //Rabbit defaults the header when deserializing the message based on the IBasicProperties.DeliveryMode
-            if (TransportIntegration.Name == TransportNames.RabbitMQClassicConventionalRoutingTopology ||
-                TransportIntegration.Name == TransportNames.RabbitMQQuorumConventionalRoutingTopology ||
-                TransportIntegration.Name == TransportNames.RabbitMQClassicDirectRoutingTopology ||
-                TransportIntegration.Name == TransportNames.RabbitMQQuorumDirectRoutingTopology)
+            if (TransportIntegration.Name.Contains("RabbitMQ"))
             {
                 Assert.AreEqual("False", context.Headers[Headers.NonDurableMessage], "Should not corrupt the non-durable header");
             }
