+++ src/ServiceControl.UnitTests/API/APIApprovals.cs
@@ -14,10 +14,8 @@
     using Particular.Approvals;
     using Particular.ServiceControl;
     using Particular.ServiceControl.Licensing;
-    using PublicApiGenerator;
     using ServiceBus.Management.Infrastructure.Settings;
     using ServiceControl.Infrastructure.WebApi;
-    using ServiceControlInstaller.Engine.Instances;
 
     [TestFixture]
     class APIApprovals
@@ -118,21 +116,6 @@
             }
 
             return t.Name;
-        }
-
-        [Test]
-        public void TransportNames()
-        {
-            //HINT: Those names are used in PowerShell scripts thus constitute a public api.
-            //Also Particular.PlatformSamples relies on it to specify the learning transport.
-            var transportNamesType = typeof(TransportNames);
-            var publicTransportNames = transportNamesType.Assembly.GeneratePublicApi(new ApiGeneratorOptions
-            {
-                IncludeTypes = new[] { transportNamesType },
-                ExcludeAttributes = new[] { "System.Reflection.AssemblyMetadataAttribute" }
-            });
-
-            Approver.Verify(publicTransportNames);
         }
 
         [Test]
