+++ src/ServiceControlInstaller.Engine.UnitTests/ApprovalFiles/InstallationTests.Audit_install_should_write_expected_config_file.RavenDB35.approved.txt
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+    <appSettings>
+        <add key="ServiceControl.Audit/Port" value="0" />
+        <add key="ServiceControl.Audit/HostName" value="localhost" />
+        <add key="ServiceControl.Audit/LogPath" value="value-not-asserted" />
+        <add key="ServiceControl.Audit/ForwardAuditMessages" value="false" />
+        <add key="ServiceControl.Audit/TransportType" value="ServiceControl.Transports.Msmq.MsmqTransportCustomization, ServiceControl.Transports.Msmq" />
+        <add key="ServiceControl.Audit/PersistenceType" value="ServiceControl.Audit.Persistence.RavenDb.RavenDbPersistenceConfiguration, ServiceControl.Audit.Persistence.RavenDb" />
+        <add key="ServiceControl.Audit/AuditRetentionPeriod" value="00:00:00" />
+        <add key="ServiceControl.Audit/EnableFullTextSearchOnBodies"
+            value="false" />
+        <add key="Raven/ESENT/MaxPageSize" value="4096" />
+        <add key="ServiceControl.Audit/RavenDB35/RunCleanupBundle" value="true" />
+        <add key="ServiceControl.Audit/DBPath" value="value-not-asserted" />
+        <add key="ServiceControl.Audit/DatabaseMaintenancePort" value="33333" />
+    </appSettings>
+    <runtime>
+        <gcServer enabled="true" />
+    </runtime>
+</configuration>
